refactor(FE): `useMintFakemon` hook
